+++ packages/backend/src/nest/connections-manager/connections-manager.service.ts
@@ -108,8 +108,12 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
 
   async onModuleInit() {
     const webcrypto = new Crypto()
-    // @ts-ignore
-    global.crypto = webcrypto
+    // https://github.com/lobehub/lobehub/issues/5315#issuecomment-2572703223
+    // TODO https://github.com/TryQuiet/quiet/issues/3123 replace if not needed
+    Object.defineProperty(global, 'crypto', {
+      value: webcrypto,
+      writable: true,
+    })
 
     setEngine(
       'newEngine',
@@ -729,9 +733,9 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
           try {
             const newInvite = this.sigChainService.getActiveChain().invites.createLongLivedUserInvite()
             const qssInitStatus = await this.qssService.getQssInitStatus()
+            // create the lockboxes using invite-based keys for users to self-assign the MEMBER role
             if (qssInitStatus.qssEnabled) {
-              const salt = uint8arrays.toString(randomBytes(32), 'hex')
-              const lbs = this.sigChainService.activeChain.lockbox.createInviteLockboxes(newInvite.seed, 'foobar')
+              this.sigChainService.activeChain.lockbox.createInviteLockboxes(newInvite.seed, newInvite.salt)
             }
             await this.sigChainService.saveChain(this.sigChainService.activeChainTeamName)
             this.serverIoProvider.io.emit(SocketEvents.CREATED_LONG_LIVED_LFA_INVITE, newInvite)