+++ packages/backend/src/nest/storage/orbitDb/orbitDb.service.ts
@@ -10,7 +10,6 @@ import { MessagesAccessController } from '../channels/messages/orbitdb/MessagesA
 import {
   createOrbitDB,
   type OrbitDBType,
-  type IdentitiesType,
   useAccessController as orbitDbUseAccessController, // this is to fix a linting issue about react hooks
   ComposedStorage,
   LRUStorage,
@@ -20,6 +19,7 @@ import {
   Entry,
   DatabaseType,
   LogType,
+  IdentitiesType,
 } from '@orbitdb/core'
 import { HeliaLibp2p } from 'helia'
 import { OrbitDbStorage } from '../../types'
@@ -34,7 +34,7 @@ import { LFAIdentities } from './identity/lfa/lfa-identity.service'
 export class OrbitDbService {
   private orbitDbInstance: OrbitDBType | undefined = undefined
   private stores: Record<string, DatabaseType> = {}
-  public identities: any | undefined = undefined
+  public identities: LFAIdentities | undefined = undefined
   public static readonly events = new EventEmitter()
 
   private readonly logger = createLogger(OrbitDbService.name)
@@ -80,7 +80,7 @@ export class OrbitDbService {
       ipfs,
       id: peerId.toString(),
       directory: this.orbitDbDir,
-      identities: this.identities,
+      identities: this.identities as any, // our type diverges from the base type
     })
 
     this.orbitDbInstance = orbitDb