+++ packages/identity/src/test/helpers.ts
@@ -45,7 +45,12 @@ export function setupCrypto() {
       subtle: webcrypto.subtle,
     })
   )
-  global.crypto = webcrypto
+  // https://github.com/lobehub/lobehub/issues/5315#issuecomment-2572703223
+  // TODO https://github.com/TryQuiet/quiet/issues/3123 replace if not needed
+  Object.defineProperty(global, 'crypto', {
+    value: webcrypto,
+    writable: true,
+  })
 }
 
 export const createRootCertificateTestHelper = async (commonName: string): Promise<RootCA> => {