+++ packages/mobile/android/app/src/main/java/com/quietmobile/Communication/CommunicationModule.java
@@ -104,9 +104,9 @@ public class CommunicationModule extends ReactContextBaseJavaModule {
     private static void deleteBackendData() {
         Context context = reactContext.getApplicationContext();
         try {
-            FileUtils.deleteDirectory(new File(context.getFilesDir(), "backend/files6"));
+            FileUtils.deleteDirectory(new File(context.getFilesDir(), "backend/files7"));
         } catch (IOException e) {
-            e.printStackTrace();
+            Log.e("CommunicationModule", e.toString());
         }
     }
 }