+++ packages/mobile/ios/Quiet/AppDelegate.h
@@ -1,12 +1,13 @@
 #import <RCTAppDelegate.h>
 #import <UIKit/UIKit.h>
 #import <Tor/Tor.h>
+#import <UserNotifications/UserNotifications.h>
 
 #import "RNNodeJsMobile.h"
 
 #import "Quiet-Swift.h"
 
-@interface AppDelegate : RCTAppDelegate
+@interface AppDelegate : RCTAppDelegate <UNUserNotificationCenterDelegate>
 
 @property uint16_t dataPort;