+++ packages/mobile/src/setupTests.tsx
@@ -7,7 +7,12 @@ import React from 'react'
 import { io } from 'socket.io-client'
 
 const webcrypto = new Crypto()
-global.crypto = webcrypto
+// https://github.com/lobehub/lobehub/issues/5315#issuecomment-2572703223
+// TODO https://github.com/TryQuiet/quiet/issues/3123 replace if not needed
+Object.defineProperty(global, 'crypto', {
+  value: webcrypto,
+  writable: true,
+})
 
 setEngine(
   'newEngine',
@@ -33,7 +38,7 @@ jest.mock('react-native-config', () => ({
   NODE_ENV: 'staging',
 }))
 
-jest.mock('redux-persist-filesystem-storage', () => { })
+jest.mock('redux-persist-filesystem-storage', () => {})
 
 jest.mock('redux-persist', () => {
   const real = jest.requireActual('redux-persist')
@@ -45,19 +50,29 @@ jest.mock('redux-persist', () => {
 
 jest.mock('react-native/Libraries/EventEmitter/NativeEventEmitter')
 
+jest.mock('react-native', () => {
+  const rn = jest.requireActual('react-native')
+  rn.NativeModules.CommunicationModule = {
+    requestNotificationPermission: jest.fn(),
+    checkNotificationPermission: jest.fn(),
+    handleIncomingEvents: jest.fn(),
+  }
+  return rn
+})
+
 jest.mock('redux-persist-filesystem-storage', () => ({
-  config: jest.fn()
+  config: jest.fn(),
 }))
 
 jest.mock('react-native-blob-util', () => ({
   fs: {
     dirs: {
-      DocumentDir: 'dir'
-    }
-  }
+      DocumentDir: 'dir',
+    },
+  },
 }))
 
-jest.mock('react-native-mathjax-html-to-svg', () => { })
+jest.mock('react-native-mathjax-html-to-svg', () => {})
 
 jest.mock('react-native-qrcode-svg', () => jest.fn())
 
@@ -65,46 +80,47 @@ jest.mock('react-native-progress', () => ({
   CircleSnail: jest.fn(),
 }))
 
-jest.mock(
-  '@ronradtke/react-native-markdown-display', () => ({
-    __esModule: true,
-    default: (props: any) => {
-      return <div>{props.children}</div>
-    },
-    MarkdownIt: jest.fn()
-  }))
+jest.mock('@ronradtke/react-native-markdown-display', () => ({
+  __esModule: true,
+  default: (props: any) => {
+    return <div>{props.children}</div>
+  },
+  MarkdownIt: jest.fn(),
+}))
 
 jest.mock('socket.io-client', () => ({
   io: jest.fn(),
 }))
 
-// Mocked because of: 
-// 
+// Mocked because of:
+//
 // "Invariant Violation: TurboModuleRegistry.getEnforcing(...): 'RNDocumentPicker'
 // could not be found. Verify that a module by this name is registered in the native binary."
-jest.mock('react-native-document-picker', () => { })
+jest.mock('react-native-document-picker', () => {})
 
 // Mocked because of:
 //
 // /Users/isla/Dev/quiet/packages/mobile/node_modules/react-native-image-picker/src/index.ts:1
 // ({"Object.<anonymous>":function(module,exports,require,__dirname,__filename,jest){import {Platform} from 'react-native';
 // ^^^^^^
-// 
+//
 // SyntaxError: Cannot use import statement outside a module
-jest.mock('react-native-image-picker', () => { })
+jest.mock('react-native-image-picker', () => {})
 
 jest.mock('react-native-device-info', () => {
   return {
-    getVersion: () => { return '1.0.0' }
+    getVersion: () => {
+      return '1.0.0'
+    },
   }
 })
 
 jest.mock('react-native-safe-area-context', () => ({
   useSafeAreaInsets: () => {
     return {
-      bottom: 0
+      bottom: 0,
     }
-  }
+  },
 }))
 
 jest.mock('react-native-fs', () => {
@@ -156,8 +172,8 @@ jest.mock('react-native-fs', () => {
 jest.mock('react-native-file-logger', () => {
   return {
     FileLogger: {
-      configure: jest.fn()
-    }
+      configure: jest.fn(),
+    },
   }
 })